requirements.txt
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
fastapi==0.109.2
uvicorn[standard]==0.27.1
langchain>=0.1.6
langchain-community>=0.0.20
langchain-core>=0.1.25,<0.2.0
langchain-google-genai>=0.0.11
llama-index>=0.10.12
llama-index-core>=0.10.12
sentence-transformers==2.4.0
open-clip-torch==2.24.0
torch==2.2.0
torchvision==0.17.0
qdrant-client==1.7.3
pymupdf==1.23.22
pillow==10.2.0
easyocr==1.7.1
unstructured[pdf]==0.12.4
pytesseract==0.3.10
ragas==0.1.5
langfuse==2.6.3
streamlit==1.31.1
google-generativeai>=0.4.1,<0.5.0
google-genai>=0.2.0
numpy==1.26.4
pandas==2.2.0
